add minimal style
